+++ PalaceAudiobookToolkit/Player/FindawayPlayer.swift
@@ -562,7 +562,9 @@ extension FindawayPlayer: FindawayPlaybackNotificationHandlerDelegate {
                 if self.shouldPauseWhenPlaybackResumes {
                     self.performPause()
                 } else {
-                    self.playbackStatePublisher.send(.started(currentChapter.position))
+                    DispatchQueue.main.async {
+                        self.playbackStatePublisher.send(.started(currentChapter.position))
+                    }
                 }
             }
             self.shouldPauseWhenPlaybackResumes = false
@@ -588,7 +590,9 @@ extension FindawayPlayer: FindawayPlaybackNotificationHandlerDelegate {
 
     func audioEnginePlaybackFailed(_ notificationHandler: FindawayPlaybackNotificationHandler, withError error: NSError?, for chapter: FAEChapterDescription) {
         guard let locationOfError = self.chapter(for: chapter)?.position else { return }
-        self.playbackStatePublisher.send(.failed(locationOfError, error))
+        DispatchQueue.main.async {
+            self.playbackStatePublisher.send(.failed(locationOfError, error))
+        }
     }
     
     func audioEngineAudiobookCompleted(_ notificationHandler: FindawayPlaybackNotificationHandler, for audiobookID: String) {
